+++ packages/crossplatform/src/index.ts
@@ -1,0 +1,4 @@
+export {
+  platformStyleTranslation,
+  type CrossPlatformStyle
+} from './helpers/platformStyleTranslation'