+++ packages/primitives/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @interunit/primitives
 
+## 0.0.3
+
+### Patch Changes
+
+- Improve inline styles by creating merged types and allowing for some style translation between web and native (more to come).
+
 ## 0.0.2
 
 ### Patch Changes